[SPARK-40981][CONNECT][PYTHON] Support session.range in Python client.

+++ connector/connect/src/main/scala/org/apache/spark/sql/connect/dsl/package.scala
@@ -180,9 +180,9 @@ package object dsl {
   object plans { // scalastyle:ignore
     implicit class DslMockRemoteSession(val session: MockRemoteSession) {
       def range(
-          start: Option[Int],
-          end: Int,
-          step: Option[Int],
+          start: Option[Long],
+          end: Long,
+          step: Option[Long],
           numPartitions: Option[Int]): Relation = {
         val range = proto.Range.newBuilder()
         if (start.isDefined) {
